refactored upcoming events into a separate component
